fix: circular dependency in stats enum
