+++ packages/api/src/analytics/dto/stats.dto.ts
@@ -9,6 +9,7 @@ import { ApiProperty } from '@nestjs/swagger';
 import { Exclude, Expose, Type } from 'class-transformer';
 import {
   IsDate,
+  IsEnum,
   IsNotEmpty,
   IsNumber,
   IsOptional,
@@ -21,7 +22,7 @@ import {
   DtoTransformerConfig,
 } from '@/utils/types/dto.types';
 
-import { StatsType } from '../entities/stats.entity';
+import { StatsType } from '../enums/stats-type.enum';
 import { IsLessThanDate } from '../validation-rules/is-less-than-date';
 
 @Exclude()
@@ -109,8 +110,8 @@ export class StatsFindDatumDto extends StatsFindDto {
   /**
    * Type for message to retrieve.
    */
-  // @IsEnum(StatsType)
-  // @IsOptional()
+  @IsEnum(StatsType)
+  @IsOptional()
   type: StatsType;
 }
