+++ packages/api/src/analytics/enums/stats-type.enum.ts
@@ -1,0 +1,15 @@
+/*
+ * Hexabot — Fair Core License (FCL-1.0-ALv2)
+ * Copyright (c) 2025 Hexastack.
+ * Full terms: see LICENSE.md.
+ */
+
+export enum StatsType {
+  outgoing = 'outgoing',
+  new_users = 'new_users',
+  all_messages = 'all_messages',
+  incoming = 'incoming',
+  returning_users = 'returning_users',
+  retention = 'retention',
+  echo = 'echo',
+}
